# go/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mormorare API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/users": {
            "get": {
                "description": "Retrieve all users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Retrieve all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/handler.UserResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorJson"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a new user",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorJson"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.ErrorJson": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handler.UserResponse": {
            "type": "object",
            "properties": {
                "birth_day": {
                    "type": "string",
                    "example": "2006-01-02"
                },
                "email": {
                    "type": "string",
                    "example": "example@test.com"
                },
                "user_name": {
                    "type": "string",
                    "example": "test1"
                }
            }
        }
    }
}